// frontend/src/types/student.ts
export interface Student {
  student_name: string;
  email: string;
  current_performance: string;
  learns_best_with: string;
  pace: string;
  complexity: string;
  strengths: string;
  needs_help_with: string;
  prior_knowledge: string;
  hobbies_interests: string;
  career_goals: string;
  example_style: string;
  question_format: string;
}